Move all the lockfile logic to utils/lockfile.go
