+++ main.go
@@ -5,14 +5,12 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"os/signal"
 	"runtime"
 	"strings"
-	"syscall"
-	"time"
 
 	"github.com/dogestry/dogestry/cli"
 	"github.com/dogestry/dogestry/config"
+	"github.com/dogestry/dogestry/utils"
 )
 
 type pullHosts []string
@@ -48,23 +46,6 @@ func init() {
 	flag.StringVar(&flLockFile, "lockfile", "", "lockfile to use while executing command, prevents parallel executions")
 }
 
-// getLock will return the lock file once it has exclusive access to it.
-// This prevents multiple processes getting a lock at the same time.
-func getLock(file string) error {
-	for {
-		_, err := os.OpenFile(file, os.O_EXCL|os.O_CREATE|os.O_WRONLY, 0666)
-		if patherr, ok := err.(*os.PathError); ok {
-			if strings.Contains(patherr.Error(), "file exists") {
-				// Lock file still exists, wait for a while and try again.
-				time.Sleep(time.Second)
-				continue
-			}
-		}
-		// Either we suceeded creating the lock or an unknown error occured.
-		return err
-	}
-}
-
 func main() {
 	runtime.GOMAXPROCS(runtime.NumCPU())
 
@@ -94,47 +75,16 @@ func main() {
 		log.Fatal(err)
 	}
 
-	signalc := make(chan os.Signal, 1)
-	signal.Notify(signalc, os.Interrupt, os.Kill, syscall.SIGTERM)
-
-	lockerrc := make(chan error)
-	locked := make(chan struct{})
 	if flLockFile != "" {
-		log.Println("Waiting for lock file")
-		go func() {
-			lockerrc <- getLock(flLockFile)
-		}()
+		utils.LockByFile(dogestryCli, args, flLockFile)
 	} else {
-		close(locked)
-	}
+		err = dogestryCli.RunCmd(args...)
 
-	errc := make(chan error)
-	go func() {
-		<-locked
-		errc <- dogestryCli.RunCmd(args...)
-	}()
-
-	for {
-		select {
-		case err := <-lockerrc:
-			if err != nil {
-				log.Println(err)
-				return
-			}
-			defer os.Remove(flLockFile)
-			close(locked)
-			// We don't expect more than one error so disable this channel
-			lockerrc = nil
-		case err := <-errc:
-			if err != nil {
-				log.Println(err)
-			}
+		if err == nil {
 			dogestryCli.Cleanup()
-			return
-		case <-signalc:
-			log.Println("Got signal, exiting")
-			return
-			// TODO: Also make it possible for dogestry cli to cancel pending actions.
+		} else {
+			dogestryCli.Cleanup()
+			log.Fatal(err)
 		}
 	}
 }
